+++ src/main/java/io/github/astrapi69/crypto/factory/PinFactory.java
@@ -22,37 +22,38 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.crypto.factories;
+package io.github.astrapi69.crypto.factory;
 
-import java.security.spec.AlgorithmParameterSpec;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.Comparator;
+import java.util.List;
 
-import javax.crypto.spec.PBEParameterSpec;
+import lombok.NonNull;
+import io.github.astrapi69.collections.list.ListFactory;
 
 /**
- * The factory class {@link AlgorithmParameterSpecFactory} holds methods for creating
- * {@link AlgorithmParameterSpec} objects.
+ * The factory class {@link PinFactory} holds methods for creating list of possible pins
  */
-public final class AlgorithmParameterSpecFactory
+public final class PinFactory
 {
-
-	private AlgorithmParameterSpecFactory()
+	private PinFactory()
 	{
 	}
 
-	/**
-	 * Factory method for creating a new {@link PBEParameterSpec} from the given salt and iteration
-	 * count.
-	 *
-	 * @param salt
-	 *            the salt
-	 * @param iterationCount
-	 *            the iteration count
-	 * @return the new {@link PBEParameterSpec} from the given salt and iteration count.
-	 */
-	public static AlgorithmParameterSpec newPBEParameterSpec(final byte[] salt,
-		final int iterationCount)
+	public static List<String> newPins(@NonNull List<LocalDate> localDates,
+		@NonNull List<String> datePatterns)
 	{
-		final AlgorithmParameterSpec paramSpec = new PBEParameterSpec(salt, iterationCount);
-		return paramSpec;
+		List<String> result = ListFactory.newArrayList();
+
+		for (LocalDate localDate : localDates)
+		{
+			for (String datePattern : datePatterns)
+			{
+				result.add(localDate.format(DateTimeFormatter.ofPattern(datePattern)));
+			}
+		}
+		result.sort(Comparator.naturalOrder());
+		return result;
 	}
 }