+++ src/main/java/io/github/astrapi69/crypto/key/reader/EncryptedPrivateKeyReader.java
@@ -54,9 +54,9 @@ import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter;
 import org.bouncycastle.openssl.jcajce.JcePEMDecryptorProviderBuilder;
 
 import io.github.astrapi69.crypto.algorithm.KeyPairGeneratorAlgorithm;
-import io.github.astrapi69.crypto.factories.CipherFactory;
-import io.github.astrapi69.crypto.factories.KeySpecFactory;
-import io.github.astrapi69.crypto.factories.SecretKeyFactoryExtensions;
+import io.github.astrapi69.crypto.factory.CipherFactory;
+import io.github.astrapi69.crypto.factory.KeySpecFactory;
+import io.github.astrapi69.crypto.factory.SecretKeyFactoryExtensions;
 import io.github.astrapi69.crypto.provider.SecurityProvider;
 
 /**